requirements.txt
# 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 데이터베이스
pymysql==1.1.0
cryptography==41.0.7

# Redis
redis==5.0.1

# HTTP 클라이언트
httpx==0.25.2

# WebSocket
websockets==12.0

# 환경 변수
python-dotenv==1.0.0

# 암호화
bcrypt==4.1.2

# JWT
pyjwt==2.8.0

# 스케줄링
schedule==1.2.0

# OpenAI
openai==1.3.7

# 데이터 처리
pydantic==2.5.0

# 로깅
structlog==23.2.0

# 유틸리티
python-multipart==0.0.6